requirements.txt
pytest
pylint
pandas
geonamescache
rapidfuzz
tabulate
boto3
python-dotenv